requirements.txt
requests 
pytest
python-dotenv